Refactored plotting + Plotting capability now exists in `Framework` + Moved Parameters to public methods `plot2D` and `plot3D` so that they can be displayed + Slightly improved docstrings + Added methods `plot_inf_flex` and `plot_stress`
