+++ pyrigi/framework.py
@@ -17,22 +17,25 @@ from __future__ import annotations
 from copy import deepcopy
 from itertools import combinations
 from random import randrange
+import distinctipy
 
 import networkx as nx
 import sympy as sp
 import numpy as np
 import functools
 
-from sympy import Matrix, flatten, binomial
+from sympy import Matrix, flatten, binomial, sympify
 
 from pyrigi.data_type import (
     Vertex,
     Edge,
     Point,
     Stress,
+    Inf_Flex,
     point_to_vector,
     Sequence,
     Coordinate,
+    DirectedEdge,
 )
 
 from pyrigi.graph import Graph
@@ -291,16 +294,24 @@ class Framework(object):
         """
         return deepcopy(self._graph)
 
-    @doc_category("Other")
+    @doc_category("Plotting")
     def _plot_with_2D_realization(
         self,
+        ax: Axes,
         realization: dict[Vertex, Point],
-        inf_flex: dict[Vertex, Sequence[Coordinate]] = None,
-        stress: dict[Edge, Coordinate] = None,
-        vertex_color="#ff8c00",
-        edge_width=1.5,
-        curved_edges=False,
-        **kwargs,
+        vertex_size: int = 300,
+        vertex_color: str = "#ff8c00",
+        vertex_shape: str = "o",
+        vertex_labels: bool = True,
+        edge_width: float = 2.5,
+        edge_color: (
+            str | Sequence[Sequence[Edge]] | dict[str : Sequence[Edge]]
+        ) = "black",
+        edge_style: str = "solid",
+        font_size: int = 12,
+        font_color: str = "whitesmoke",
+        curved_edges: bool = False,
+        connection_style: float | Sequence[float] | dict[Edge, float] = np.pi / 6,
     ) -> None:
         """
         Plot the graph of the framework with the given realization in the plane.
@@ -323,55 +334,98 @@ class Framework(object):
             Optional parameter for plotting an equilibrium stress. We expect
             it to have the format `Dict[Edge, Coordinate]`.
         """
+        edge_color_array, edge_list_ref = self._resolve_edge_colors(edge_color)
 
-        self._graph.plot(
-            placement=realization,
-            inf_flex=inf_flex,
-            stress=stress,
-            vertex_color=vertex_color,
-            edge_width=edge_width,
-            curved_edges=curved_edges,
-            **kwargs,
-        )
+        if not curved_edges:
+            nx.draw(
+                self._graph,
+                pos=realization,
+                ax=ax,
+                node_size=vertex_size,
+                node_color=vertex_color,
+                node_shape=vertex_shape,
+                with_labels=vertex_labels,
+                width=edge_width,
+                edge_color=edge_color_array,
+                font_color=font_color,
+                font_size=font_size,
+                edgelist=edge_list_ref,
+                style=edge_style,
+            )
+        else:
+            newGraph = nx.MultiDiGraph()
+            connection_style = self._resolve_connection_style(connection_style)
+            for e, style in connection_style.items():
+                newGraph.add_edge(e[0], e[1], weight=style)
+            plt.box(False)  # Manually removes the frame of the plot
+            plt.tick_params(
+                left=False,
+                right=False,
+                labelleft=False,
+                labelbottom=False,
+                bottom=False,
+            )  # Removes the ticks
+            nx.draw_networkx_nodes(
+                newGraph,
+                realization,
+                ax=ax,
+                node_size=vertex_size,
+                node_color=vertex_color,
+                node_shape=vertex_shape,
+            )
+            nx.draw_networkx_labels(
+                newGraph, realization, ax=ax, font_color=font_color, font_size=font_size
+            )
+            for edge in newGraph.edges(data=True):
+                nx.draw_networkx_edges(
+                    newGraph,
+                    realization,
+                    ax=ax,
+                    width=edge_width,
+                    edge_color=edge_color_array,
+                    arrows=True,
+                    arrowstyle="-",
+                    edgelist=[(edge[0], edge[1])],
+                    connectionstyle=f"Arc3, rad = {edge[2]['weight']}",
+                )
 
-    @doc_category("Other")
-    def _plot_using_projection_matrix(
-        self,
-        projection_matrix: Matrix,
-        **kwargs,
-    ) -> None:
-        """
-        Plot the framework with the realization projected using the given matrix.
-
-        For description of other parameters see :meth:`.Framework.plot`.
-
-        Parameters
-        ----------
-        projection_matrix:
-            The matrix used for projection.
-            The matrix must have dimensions ``(2, dim)``,
-            where ``dim`` is the dimension of the framework.
-        """
-
-        placement = {}
-        for vertex, position in self.realization(
-            as_points=False, numerical=True
-        ).items():
-            placement[vertex] = np.dot(projection_matrix, np.array(position))
-
-        self._plot_with_2D_realization(placement, **kwargs)
-
-    @doc_category("Other")
+    @doc_category("Plotting")
     def plot2D(  # noqa: C901
         self,
-        coordinates: Sequence[int] = None,
-        inf_flex: Matrix | int | dict[Vertex, Sequence[Coordinate]] = None,
-        stress: Matrix | int | dict[Edge, Coordinate] = None,
         projection_matrix: Matrix = None,
-        return_matrix: bool = False,
         random_seed: int = None,
+        coordinates: Sequence[int] = None,
+        inf_flex: int | Inf_Flex = None,
+        stress: int | Stress = None,
+        return_matrix: bool = False,
+        vertex_size: int = 300,
+        vertex_color: str = "#ff8c00",
+        vertex_shape: str = "o",
+        vertex_labels: bool = True,
+        edge_width: float = 2.5,
+        edge_color: (
+            str | Sequence[Sequence[Edge]] | dict[str : Sequence[Edge]]
+        ) = "black",
+        edge_style: str = "solid",
+        flex_width: float = 1.5,
+        flex_length: float = 0.15,
+        flex_color: (
+            str | Sequence[Sequence[Edge]] | dict[str : Sequence[Edge]]
+        ) = "limegreen",
+        flex_style: str = "solid",
+        flex_arrowsize: int = 20,
+        stress_color: str = "orangered",
+        stress_fontsize: int = 10,
+        stress_label_pos: float | dict[DirectedEdge, float] = 0.5,
+        stress_rotate_labels: bool = True,
+        stress_normalization: bool = False,
+        font_size: int = 12,
+        font_color: str = "whitesmoke",
+        canvas_width: float = 6.4,
+        canvas_height: float = 4.8,
+        aspect_ratio: float = 1.0,
         curved_edges: bool = False,
-        **kwargs,
+        connection_style: float | Sequence[float] | dict[Edge, float] = np.pi / 6,
     ) -> Optional[Matrix]:
         """
         Plot this framework in 2D.
@@ -420,63 +474,119 @@ class Framework(object):
             (i.e. coordinates).
         return_matrix:
             If True the matrix used for projection into 2D is returned.
+        vertex_size:
+            The size of the vertices.
+        vertex_color:
+            The color of the vertices. The color can be a string or an rgb (or rgba)
+            tuple of floats from 0-1.
+        vertex_shape:
+            The shape of the vertices specified as as matplotlib.scatter
+            marker, one of ``so^>v<dph8``.
+        vertex_labels:
+            If ``True`` (default), vertex labels are displayed.
+        edge_width:
+        edge_color:
+            If a single color is given as a string or rgb (or rgba) tuple
+            of floats from 0-1, then all edges get this color.
+            If a (possibly incomplete) partition of the edges is given,
+            then each part gets a different color.
+            If a dictionary from colors to a list of edge is given,
+            edges are colored accordingly.
+            The edges missing in the partition/dictionary, are colored black.
+        edge_style:
+            Edge line style: ``-``/``solid``, ``--``/``dashed``,
+            ``-.``/``dashdot`` or ``:``/``dotted``. By default '-'.
+        flex_width:
+            The width of the infinitesimal flex's arrow tail.
+        flex_color:
+            The color of the infinitesimal flex is by default 'limegreen'.
+        flex_style:
+            Line Style: ``-``/``solid``, ``--``/``dashed``,
+            ``-.``/``dashdot`` or ``:``/``dotted``. By default '-'.
+        flex_length:
+            Length of the displayed flex relative to the total canvas
+            diagonal in percent. By default 15%.
+        flex_arrowsize:
+            Size of the arrowhead's length and width.
+        stress_color:
+            Color of the font used to label the edges with stresses.
+        stress_fontsize:
+            Fontsize of the stress labels.
+        stress_label_pos:
+            Position of the stress label along the edge. `float` numbers
+            from the interval `[0,1]` are allowed. `0` represents the head
+            of the edge, `0.5` the center and `1` the edge's tail. The position
+            can either be specified for all edges equally or as a
+            `dict[Edge, float]` of ordered edges. Omitted edges are set to `0.5`.
+        stress_rotate_labels:
+            A boolean indicating whether the stress label should be rotated.
+        stress_normalization:
+            A boolean indicating whether the stress values should be turned into
+            floating point numbers. If ``True``, the stress is automatically normalized.
+        font_size:
+            The size of the font used for the labels.
+        font_color:
+            The color of the font used for the labels.
+        canvas_width:
+            The width of the canvas in inches.
+        canvas_height:
+            The height of the canvas in inches.
+        aspect_ratio:
+            The ratio of y-unit to x-unit. By default 1.0.
         curved_edges:
-            If the points are collinear (or close to collinear), then we can set this
-            ``bool`` to ``True``. The result is that the edges are plotted as arcs.
+            If the edges are too close to each other, we can decide to
+            visualize them as arcs.
+        connection_style:
+            In case of curvilinear plotting (``curved_edges=True``), the edges
+            are displayed as arcs. With this parameter, we can set the
+            pitch of these arcs and it is in radians. It can either be
+            specified for each arc (``connection_style=0.5``) or individually
+            as a ``list`` and ``dict``
+            (``connection_style={(0,1):0.5, (1,2):-0.5}``). It is possible to
+            provide fewer edges when the input is a ``dict``; the remaining
+            edges are padded with zeros in that case.
 
-        TODO
-        -----
-        project the inf-flex as well in `_plot_using_projection_matrix`.
+        Examples
+        --------
+        >>> from pyrigi import frameworkDB
+        >>> F = frameworkDB.Complete(4, dim=1)
+        >>> F.plot2D(stress=0);
+        >>> F = frameworkDB.Frustum(3)
+        >>> F.plot2D(inf_flex=0, stress=0);
         """
-        inf_flex_pointwise = None
-        if inf_flex is not None:
-            if isinstance(inf_flex, int) and inf_flex >= 0:
-                inf_flex_basis = self.nontrivial_inf_flexes()
-                if inf_flex >= len(inf_flex_basis):
-                    raise IndexError(
-                        "The value of `inf_flex` exceeds "
-                        + "the dimension of the space "
-                        + "of infinitesimal flexes."
-                    )
-                inf_flex_pointwise = self._transform_inf_flex_to_pointwise(
-                    inf_flex_basis[inf_flex]
-                )
-            elif isinstance(inf_flex, Matrix):
-                inf_flex_pointwise = self._transform_inf_flex_to_pointwise(inf_flex)
-            elif isinstance(inf_flex, dict) and all(
-                isinstance(inf_flex[key], list | tuple) for key in inf_flex.keys()
-            ):
-                inf_flex_pointwise = inf_flex
-            else:
-                raise TypeError("`inf_flex` does not have the correct Type.")
+        plotting_args = {
+            "vertex_size": vertex_size,
+            "vertex_color": vertex_color,
+            "vertex_shape": vertex_shape,
+            "vertex_labels": vertex_labels,
+            "edge_width": edge_width,
+            "edge_color": edge_color,
+            "edge_style": edge_style,
+            "font_size": font_size,
+            "font_color": font_color,
+            "connection_style": connection_style,
+        }
+        flex_args = {
+            "flex_width": flex_width,
+            "flex_length": flex_length,
+            "flex_color": flex_color,
+            "flex_style": flex_style,
+            "flex_arrowsize": flex_arrowsize,
+        }
+        stress_args = {
+            "stress_color": stress_color,
+            "stress_fontsize": stress_fontsize,
+            "stress_label_pos": stress_label_pos,
+            "stress_rotate_labels": stress_rotate_labels,
+            "stress_normalization": stress_normalization,
+            "connection_style": connection_style,
+        }
 
-            if not self.is_dict_inf_flex(inf_flex_pointwise):
-                raise ValueError(
-                    "The provided `inf_flex` is not an infinitesimal flex."
-                )
-
-        stress_edgewise = None
-        if stress is not None:
-            if isinstance(stress, int) and stress >= 0:
-                stresses = self.stresses()
-                if stress >= len(stresses):
-                    raise IndexError(
-                        "The value of `stress` exceeds "
-                        + "the dimension of the space "
-                        + "of equilibrium stresses."
-                    )
-                stress_edgewise = self._transform_stress_to_edgewise(stresses[stress])
-            elif isinstance(stress, Matrix):
-                stress_edgewise = self._transform_stress_to_edgewise(stress)
-            elif isinstance(stress, dict) and all(
-                isinstance(stress[key], int | float | str) for key in stress.keys()
-            ):
-                stress_edgewise = stress
-            else:
-                raise TypeError("`stress` does not have the correct Type.")
-
-            if not self.is_dict_stress(stress_edgewise):
-                raise ValueError("The provided `stress` is not an equilibrium stress.")
+        fig, ax = plt.subplots()
+        ax.set_adjustable("datalim")
+        fig.set_figwidth(canvas_width)
+        fig.set_figheight(canvas_height)
+        ax.set_aspect(aspect_ratio)
 
         if self._dim == 1:
             placement = {}
@@ -485,28 +595,27 @@ class Framework(object):
             ).items():
                 placement[vertex] = np.append(np.array(position), 0)
 
-            if inf_flex_pointwise is not None:
-                inf_flex_pointwise = {
-                    v: (flex_v[0], 0) for v, flex_v in inf_flex_pointwise.items()
-                }
             self._plot_with_2D_realization(
-                placement,
-                inf_flex=inf_flex_pointwise,
-                stress=stress_edgewise,
-                curved_edges=True,
-                **kwargs,
+                ax, placement, curved_edges=True, **plotting_args
             )
+
+            if inf_flex is not None:
+                self._plot_inf_flex(ax, inf_flex, points=placement, **flex_args)
+            if stress is not None:
+                self._plot_stress(
+                    ax, stress, points=placement, curved_edges=True, **stress_args
+                )
             return
 
+        placement = self.realization(as_points=True, numerical=True)
         if self._dim == 2:
-            placement = self.realization(as_points=True, numerical=True)
             self._plot_with_2D_realization(
-                placement,
-                inf_flex=inf_flex_pointwise,
-                stress=stress_edgewise,
-                curved_edges=curved_edges,
-                **kwargs,
+                ax, placement, curved_edges=curved_edges, **plotting_args
             )
+            if inf_flex is not None:
+                self._plot_inf_flex(ax, inf_flex, **flex_args)
+            if stress is not None:
+                self._plot_stress(ax, stress, curved_edges=curved_edges, **stress_args)
             return
 
         # dim > 2 -> use projection to 2D
@@ -541,12 +650,37 @@ class Framework(object):
                 self._dim, random_seed=random_seed
             )
             projection_matrix = projection_matrix.T
-        self._plot_using_projection_matrix(projection_matrix, **kwargs)
+
+        for vertex, position in self.realization(
+            as_points=False, numerical=True
+        ).items():
+            placement[vertex] = np.dot(projection_matrix, np.array(position))
+
+        self._plot_with_2D_realization(
+            ax, placement, curved_edges=curved_edges, **plotting_args
+        )
+        if inf_flex is not None:
+            self._plot_inf_flex(
+                ax,
+                inf_flex,
+                points=placement,
+                projection_matrix=projection_matrix,
+                **flex_args,
+            )
+        if stress is not None:
+            self._plot_stress(
+                ax,
+                stress,
+                points=placement,
+                projection_matrix=projection_matrix,
+                curved_edges=curved_edges,
+                **stress_args,
+            )
         if return_matrix:
             return projection_matrix
 
-    @doc_category("Other")
-    def _animate_rotation_around_axis(
+    @doc_category("Plotting")
+    def animate3D(
         self,
         vertex_color: str = "#ff8c00",
         vertex_shape: str = "o",
@@ -586,7 +720,7 @@ class Framework(object):
         --------
         >>> from pyrigi import frameworkDB
         >>> F = frameworkDB.Complete(4, dim=3)
-        >>> F._animate_rotation_around_axis();
+        >>> F.animate3D();
         """
         # Creation of the figure
         fig = plt.figure()
@@ -594,7 +728,7 @@ class Framework(object):
         ax.grid(False)
         ax.set_axis_off()
 
-        edge_color_array, edge_list_ref = self._graph._resolve_edge_colors(edge_color)
+        edge_color_array, edge_list_ref = self._resolve_edge_colors(edge_color)
 
         # Limits of the axes
         abs_list = [list(abs(i)) for i in self._realization.values()]
@@ -738,15 +872,40 @@ class Framework(object):
             plt.show()
             return
 
-    @doc_category("Other")
+    @doc_category("Plotting")
     def plot3D(
         self,
-        coordinates: Sequence[int] = None,
         projection_matrix: Matrix = None,
-        return_matrix: bool = False,
         random_seed: int = None,
-        animation: bool = False,
-        **kwargs,
+        coordinates: Sequence[int] = None,
+        inf_flex: int | Inf_Flex = None,
+        stress: int | Stress = None,
+        return_matrix: bool = False,
+        vertex_size: int = 200,
+        vertex_color: str = "#ff8c00",
+        vertex_shape: str = "o",
+        vertex_labels: bool = True,
+        edge_width: float = 2.5,
+        edge_color: (
+            str | Sequence[Sequence[Edge]] | dict[str : Sequence[Edge]]
+        ) = "black",
+        edge_style: str = "solid",
+        flex_width: float = 2.5,
+        flex_length: float = 0.75,
+        flex_color: (
+            str | Sequence[Sequence[Edge]] | dict[str : Sequence[Edge]]
+        ) = "limegreen",
+        flex_style: str = "solid",
+        flex_arrowsize: int = 20,
+        stress_color: str = "orangered",
+        stress_fontsize: int = 8,
+        stress_label_pos: float | dict[DirectedEdge, float] = 0.5,
+        stress_rotate_labels: bool = True,
+        stress_normalization: bool = False,
+        font_size: int = 12,
+        font_color: str = "whitesmoke",
+        equal_aspect_ratio: bool = True,
+        padding: float = 0.01,
     ) -> Optional[Matrix]:
         """
         Plot the provided framework in 3D.
@@ -773,33 +932,69 @@ class Framework(object):
             Indices of three coordinates to which the framework is projected.
         return_matrix:
             If ``True``, the matrix used for projection into 3D is returned.
-        animation:
-            If ``True``, the plot is a rotating figure.
 
         Notes
         -----
         See :meth:`.Framework._plot_using_projection_matrix_3D` for a full
         list of parameters.
 
-        TODO
-        ----
-        Project the `inf_flex` as well.
-
         Examples
         --------
         >>> from pyrigi import frameworkDB
-        >>> F = frameworkDB.Complete(4, dim=3)
-        >>> F.plot3D();
+        >>> F = frameworkDB.Octahedron(realization="Bricard_plane")
+        >>> F.plot3D(inf_flex=0, stress=0);
         """
+        plotting_args = {
+            "vertex_size": vertex_size,
+            "vertex_color": vertex_color,
+            "vertex_shape": vertex_shape,
+            "vertex_labels": vertex_labels,
+            "edge_width": edge_width,
+            "edge_color": edge_color,
+            "edge_style": edge_style,
+            "font_size": font_size,
+            "font_color": font_color,
+        }
+        flex_args = {
+            "flex_width": flex_width,
+            "flex_length": flex_length,
+            "flex_color": flex_color,
+            "flex_style": flex_style,
+            "flex_arrowsize": flex_arrowsize,
+        }
+        stress_args = {
+            "stress_color": stress_color,
+            "stress_fontsize": stress_fontsize,
+            "stress_label_pos": stress_label_pos,
+            "stress_rotate_labels": stress_rotate_labels,
+            "stress_normalization": stress_normalization,
+        }
 
         if self._dim == 1 or self._dim == 2:
-            return self.plot2D(**kwargs)
+            return self.plot2D(
+                inf_flex=inf_flex,
+                stress=stress,
+                **(plotting_args | flex_args | stress_args),
+            )
 
-        if self._dim == 3 and not animation:
-            return self._plot_with_3D_realization(**kwargs)
-
-        elif self._dim == 3 and animation:
-            return self._animate_rotation_around_axis(**kwargs)
+        fig = plt.figure(dpi=150)
+        ax = fig.add_subplot(111, projection="3d")
+        ax.grid(False)
+        ax.set_axis_off()
+        placement = self.realization(as_points=True, numerical=True)
+        if self._dim == 3:
+            self._plot_with_3D_realization(
+                ax,
+                placement,
+                equal_aspect_ratio=equal_aspect_ratio,
+                padding=padding,
+                **plotting_args,
+            )
+            if inf_flex is not None:
+                self._plot_inf_flex(ax, inf_flex, points=placement, **flex_args)
+            if stress is not None:
+                self._plot_stress(ax, stress, points=placement, **stress_args)
+            return
 
         # dim > 3 -> use projection to 3D
         if coordinates is not None:
@@ -834,21 +1029,48 @@ class Framework(object):
                 self._dim, random_seed=random_seed
             )
             projection_matrix = projection_matrix.T
-        self._plot_with_3D_realization(projection_matrix=projection_matrix, **kwargs)
+        for vertex, position in self.realization(
+            as_points=False, numerical=True
+        ).items():
+            placement[vertex] = np.dot(projection_matrix, np.array(position))
+
+        self._plot_with_3D_realization(
+            ax,
+            placement,
+            equal_aspect_ratio=equal_aspect_ratio,
+            padding=padding,
+            **plotting_args,
+        )
+        if inf_flex is not None:
+            self._plot_inf_flex(
+                ax,
+                inf_flex,
+                points=placement,
+                projection_matrix=projection_matrix,
+                **flex_args,
+            )
+        if stress is not None:
+            self._plot_stress(
+                ax,
+                stress,
+                points=placement,
+                projection_matrix=projection_matrix,
+                **stress_args,
+            )
         if return_matrix:
             return projection_matrix
 
-    @doc_category("Other")
+    @doc_category("Plotting")
     def _plot_with_3D_realization(
         self,
-        inf_flex: Matrix | int | dict[Vertex, Sequence[Coordinate]] = None,
-        projection_matrix: Matrix = None,
+        ax: Axes,
+        realization: dict[Vertex, Point],
         vertex_color: str = "#ff8c00",
         vertex_size: int = 200,
         vertex_shape: str = "o",
         vertex_labels: bool = True,
         font_color: str = "whitesmoke",
-        fontsize: int = 10,
+        font_size: int = 10,
         edge_width: float = 2.5,
         edge_color: (
             str | Sequence[Sequence[Edge]] | dict[str : Sequence[Edge]]
@@ -856,7 +1078,6 @@ class Framework(object):
         edge_style: str = "solid",
         equal_aspect_ratio: bool = True,
         padding: float = 0.01,
-        **kwargs,
     ) -> None:
         """
         Plot the graph of the framework with the given realization in the plane.
@@ -887,7 +1108,7 @@ class Framework(object):
             marker, one of ``so^>v<dph8``.
         vertex_labels:
             If ``True`` (default), vertex labels are displayed.
-        fontsize:
+        font_size:
             The size of the font used for the labels.
         font_color:
             The color of the font used for the labels.
@@ -914,34 +1135,15 @@ class Framework(object):
         -----
         The parameters for `inf_flex`-plotting are listed in
         the API reference.
-
-        Examples
-        --------
-        >>> from pyrigi import frameworkDB
-        >>> F = frameworkDB.Complete(4, dim=3)
-        >>> F._plot_with_3D_realization();
         """
         # Create a figure for the rapresentation of the framework
-        fig = plt.figure(dpi=150)
-        ax = fig.add_subplot(111, projection="3d")
-        ax.grid(False)
-        ax.set_axis_off()
 
-        edge_color_array, edge_list_ref = self._graph._resolve_edge_colors(edge_color)
-
-        if projection_matrix is None:
-            pos = self.realization(as_points=True, numerical=True)
-        else:
-            pos = {}
-            for vertex, position in self.realization(
-                as_points=False, numerical=True
-            ).items():
-                pos[vertex] = np.dot(projection_matrix, np.array(position))
+        edge_color_array, edge_list_ref = self._resolve_edge_colors(edge_color)
 
         # Draw the vertices as points in the 3D enviroment
-        x_nodes = [pos[node][0] for node in self._graph.nodes]
-        y_nodes = [pos[node][1] for node in self._graph.nodes]
-        z_nodes = [pos[node][2] for node in self._graph.nodes]
+        x_nodes = [realization[node][0] for node in self._graph.nodes]
+        y_nodes = [realization[node][1] for node in self._graph.nodes]
+        z_nodes = [realization[node][2] for node in self._graph.nodes]
         ax.scatter(
             x_nodes,
             y_nodes,
@@ -963,12 +1165,12 @@ class Framework(object):
 
         for i in range(len(edge_list_ref)):
             edge = edge_list_ref[i]
-            x = [pos[edge[0]][0], pos[edge[1]][0]]
-            y = [pos[edge[0]][1], pos[edge[1]][1]]
-            z = [pos[edge[0]][2], pos[edge[1]][2]]
+            x = [realization[edge[0]][0], realization[edge[1]][0]]
+            y = [realization[edge[0]][1], realization[edge[1]][1]]
+            z = [realization[edge[0]][2], realization[edge[1]][2]]
             ax.plot(x, y, z, c=edge_color_array[i], lw=edge_width, linestyle=edge_style)
         for node in self._graph.nodes:
-            x, y, z, *others = pos[node]
+            x, y, z, *others = realization[node]
             # To show the name of the vertex
             if vertex_labels:
                 ax.text(
@@ -977,21 +1179,16 @@ class Framework(object):
                     z,
                     str(node),
                     color=font_color,
-                    fontsize=fontsize,
+                    fontsize=font_size,
                     ha="center",
                     va="center",
-                    alpha=1,
                 )
-        self._plot_inf_flex(ax, inf_flex, **kwargs)
 
-        plt.tight_layout()
-        plt.show()
-
-    @doc_category("Other")
+    @doc_category("Plotting")
     def _plot_inf_flex(  # noqa: C901
         self,
         ax: Axes,
-        inf_flex: Matrix | int | dict[Vertex, Sequence[Coordinate]],
+        inf_flex: Matrix | Inf_Flex,
         points: dict[Vertex, Point] = None,
         flex_width: float = 2.5,
         flex_length: float = 0.65,
@@ -999,6 +1196,9 @@ class Framework(object):
             str | Sequence[Sequence[Edge]] | dict[str : Sequence[Edge]]
         ) = "limegreen",
         flex_style: str = "solid",
+        flex_arrowsize: int = 20,
+        projection_matrix: Matrix = None,
+        **kwargs,
     ) -> None:
         """
         Adds infinitesimal flexes based in the `points` as vectors to the axis `ax`.
@@ -1024,107 +1224,304 @@ class Framework(object):
         flex_style:
             Line Style: ``-``/``solid``, ``--``/``dashed``,
             ``-.``/``dashdot`` or ``:``/``dotted``. By default '-'.
+        projection_matrix:
         """
-        if inf_flex is not None:
-            inf_flex_pointwise = None
-            if isinstance(inf_flex, int) and inf_flex >= 0:
-                inf_flex_basis = self.nontrivial_inf_flexes()
-                if inf_flex >= len(inf_flex_basis):
-                    raise IndexError(
-                        "The value of inf_flex exceeds "
-                        + "the dimension of the space "
-                        + "of infinitesimal flexes."
-                    )
-                inf_flex_pointwise = self._transform_inf_flex_to_pointwise(
-                    inf_flex_basis[inf_flex]
+        inf_flex_pointwise = None
+        if isinstance(inf_flex, int) and inf_flex >= 0:
+            inf_flex_basis = self.nontrivial_inf_flexes()
+            if inf_flex >= len(inf_flex_basis):
+                raise IndexError(
+                    "The value of inf_flex exceeds "
+                    + "the dimension of the space "
+                    + "of infinitesimal flexes."
                 )
-            elif isinstance(inf_flex, Matrix):
-                inf_flex_pointwise = self._transform_inf_flex_to_pointwise(inf_flex)
-            elif isinstance(inf_flex, dict) and all(
-                isinstance(inf_flex[key], Sequence) for key in inf_flex.keys()
-            ):
-                inf_flex_pointwise = inf_flex
-            else:
-                raise TypeError("inf_flex does not have the correct Type.")
+            inf_flex_pointwise = self._transform_inf_flex_to_pointwise(
+                inf_flex_basis[inf_flex]
+            )
+        elif isinstance(inf_flex, Matrix):
+            inf_flex_pointwise = self._transform_inf_flex_to_pointwise(inf_flex)
+        elif isinstance(inf_flex, dict) and all(
+            isinstance(inf_flex[key], Sequence) for key in inf_flex.keys()
+        ):
+            inf_flex_pointwise = inf_flex
+        else:
+            raise TypeError(
+                "inf_flex does not have the correct Type or the `int` is too large."
+            )
 
-            if not self.is_dict_inf_flex(inf_flex_pointwise):
-                raise ValueError(
-                    "The provided `inf_flex` is not an infinitesimal flex."
-                )
-
-            if points is None:
-                points = self.realization(as_points=True, numerical=True)
-            elif not isinstance(points, dict):
-                raise TypeError("Realization has the wrong type!")
-            elif not all(
-                [
-                    len(points[v]) == len(points[points.keys()[0]])
-                    and len(points[v]) in [2, 3]
-                    for v in self._graph.nodes
-                ]
-            ):
-                raise ValueError(
-                    "Not all values in the realization have the same"
-                    + "length and the dimension needs to be 2 or 3."
-                )
-
-            magnidutes = []
-            for flex_key in inf_flex_pointwise.keys():
-                if len(inf_flex_pointwise[flex_key]) != len(
-                    points[list(points.keys())[0]]
-                ):
-                    raise ValueError(
-                        "The infinitesimal flex needs to be "
-                        + f"in dimension {len(points[list(points.keys())[0]])}."
-                    )
-                inf_flex = [float(x) for x in inf_flex_pointwise[flex_key]]
-                magnidutes.append(np.linalg.norm(inf_flex))
-
-            # normalize the edge lengths by the Euclidean norm of the longest one
-            flex_mag = max(magnidutes)
-            for flex_key in inf_flex_pointwise.keys():
-                if not all(entry == 0 for entry in inf_flex_pointwise[flex_key]):
-                    inf_flex_pointwise[flex_key] = tuple(
-                        flex / flex_mag for flex in inf_flex_pointwise[flex_key]
-                    )
-            # Delete the edges with zero length
+        if not self.is_dict_inf_flex(inf_flex_pointwise):
+            raise ValueError("The provided `inf_flex` is not an infinitesimal flex.")
+        if self._dim == 1:
             inf_flex_pointwise = {
-                flex_key: np.array(inf_flex_pointwise[flex_key], dtype=float)
-                for flex_key in inf_flex_pointwise.keys()
-                if not all(entry == 0 for entry in inf_flex_pointwise[flex_key])
+                v: [inf_flex_pointwise[v], 0] for v in inf_flex_pointwise.keys()
+            }
+        if projection_matrix is not None:
+            inf_flex_pointwise = {
+                v: np.dot(projection_matrix, np.array(flex))
+                for v, flex in inf_flex_pointwise.items()
             }
 
+        if points is None:
+            points = self.realization(as_points=True, numerical=True)
+        elif not isinstance(points, dict):
+            raise TypeError("Realization has the wrong type!")
+        elif not all(
+            [
+                len(points[v]) == len(points[list(points.keys())[0]])
+                and len(points[v]) in [2, 3]
+                for v in self._graph.nodes
+            ]
+        ):
+            raise ValueError(
+                "Not all values in the realization have the same"
+                + "length and the dimension needs to be 2 or 3."
+            )
+
+        magnidutes = []
+        for flex_key in inf_flex_pointwise.keys():
+            if len(inf_flex_pointwise[flex_key]) != len(points[list(points.keys())[0]]):
+                raise ValueError(
+                    "The infinitesimal flex needs to be "
+                    + f"in dimension {len(points[list(points.keys())[0]])}."
+                )
+            inf_flex = [float(x) for x in inf_flex_pointwise[flex_key]]
+            magnidutes.append(np.linalg.norm(inf_flex))
+
+        # normalize the edge lengths by the Euclidean norm of the longest one
+        flex_mag = max(magnidutes)
+        for flex_key in inf_flex_pointwise.keys():
+            if not all(entry == 0 for entry in inf_flex_pointwise[flex_key]):
+                inf_flex_pointwise[flex_key] = tuple(
+                    flex / flex_mag for flex in inf_flex_pointwise[flex_key]
+                )
+        # Delete the edges with zero length
+        inf_flex_pointwise = {
+            flex_key: np.array(inf_flex_pointwise[flex_key], dtype=float)
+            for flex_key in inf_flex_pointwise.keys()
+            if not all(entry == 0 for entry in inf_flex_pointwise[flex_key])
+        }
+
+        if len(inf_flex_pointwise[list(inf_flex_pointwise.keys())[0]]) == 2:
+            x_canvas_width = ax.get_xlim()[1] - ax.get_xlim()[0]
+            y_canvas_width = ax.get_ylim()[1] - ax.get_ylim()[0]
+            arrow_length = np.sqrt(x_canvas_width**2 + y_canvas_width**2) * flex_length
+            H = nx.DiGraph([(v, str(v) + "_flex") for v in inf_flex_pointwise.keys()])
+            H_placement = {
+                str(v)
+                + "_flex": np.array(
+                    [
+                        points[v][0] + arrow_length * inf_flex_pointwise[v][0],
+                        points[v][1] + arrow_length * inf_flex_pointwise[v][1],
+                    ],
+                    dtype=float,
+                )
+                for v in inf_flex_pointwise.keys()
+            }
+            H_placement.update(
+                {v: np.array(points[v], dtype=float) for v in inf_flex_pointwise.keys()}
+            )
+            if (
+                not isinstance(flex_color, str | list)
+                or isinstance(flex_color, list)
+                and not len(flex_color) == len(inf_flex_pointwise)
+            ):
+                raise TypeError(
+                    "`flex_color` must either be a `str` specifying"
+                    + "a color or a list of strings with the same"
+                    + "number as the nonzero flexes."
+                )
+            nx.draw(
+                H,
+                pos=H_placement,
+                ax=ax,
+                arrows=True,
+                arrowsize=flex_arrowsize,
+                node_size=0,
+                node_color="white",
+                width=flex_width,
+                edge_color=flex_color,
+                style=flex_style,
+                **kwargs,
+            )
+        elif self._dim == 3:
             for v in inf_flex_pointwise.keys():
-                if len(points[v]) == 2:
-                    ax.quiver(
-                        points[v][0],
-                        points[v][1],
-                        inf_flex_pointwise[v][0],
-                        inf_flex_pointwise[v][1],
-                        color=flex_color,
-                        lw=flex_width,
-                        linestyle=flex_style,
-                        length=flex_length,
-                        normalize=True,
-                        arrow_length_ratio=0.25,
-                    )
+                ax.quiver(
+                    points[v][0],
+                    points[v][1],
+                    points[v][2],
+                    inf_flex_pointwise[v][0],
+                    inf_flex_pointwise[v][1],
+                    inf_flex_pointwise[v][2],
+                    color=flex_color,
+                    lw=flex_width,
+                    linestyle=flex_style,
+                    length=flex_length,
+                    arrow_length_ratio=0.25,
+                )
+        else:
+            raise ValueError(
+                "The dimension of the infinitesimal flex needs to be between 1 and 3."
+            )
 
-                elif self._dim == 3:
-                    ax.quiver(
-                        points[v][0],
-                        points[v][1],
-                        points[v][2],
-                        inf_flex_pointwise[v][0],
-                        inf_flex_pointwise[v][1],
-                        inf_flex_pointwise[v][2],
-                        color=flex_color,
-                        lw=flex_width,
-                        linestyle=flex_style,
-                        length=flex_length,
-                        arrow_length_ratio=0.25,
-                    )
+    @doc_category("Plotting")
+    def _plot_stress(  # noqa: C901
+        self,
+        ax: Axes,
+        stress: Matrix | Stress,
+        points: dict[Vertex, Point] = None,
+        stress_color: str = "orangered",
+        stress_fontsize: int = 8,
+        stress_label_pos: float | dict[DirectedEdge, float] = 0.5,
+        stress_rotate_labels: bool = True,
+        stress_normalization: bool = False,
+        connection_style: float | dict[DirectedEdge, float] = 0.5,
+        curved_edges: bool = False,
+        **kwargs,
+    ) -> None:
+        """
+        ax:
+        stress:
+            Optional parameter for plotting a given equilibrium stress. The standard
+            input format is a ``Matrix`` that is the output of e.g. the method
+            ``Framework.stresses``. Alternatively, an ``int`` can be specified
+            to directly choose the 0,1,2,...-th equilibrium stress (according
+            to the method ``Framework.stresses``) for plotting.
+            For these input types, is important to use the same edge order as the one
+            from :meth:`.Graph.edge_list`.
+            If the edge order needs to be specified, a ``Dict[Edge, Coordinate]``
+            can be provided, which maps the edges to numbers
+            (i.e. coordinates).
+        points:
+            It is possible to provide an alternative realization.
+        stress_color:
+            Color of the font used to label the edges with stresses.
+        stress_fontsize:
+            Fontsize of the stress labels.
+        stress_label_pos:
+            Position of the stress label along the edge. `float` numbers
+            from the interval `[0,1]` are allowed. `0` represents the head
+            of the edge, `0.5` the center and `1` the edge's tail. The position
+            can either be specified for all edges equally or as a
+            `dict[Edge, float]` of ordered edges. Omitted edges are set to `0.5`.
+        stress_rotate_labels:
+            A boolean indicating whether the stress label should be rotated.
+        stress_normalization:
+            A boolean indicating whether the stress values should be turned into
+            floating point numbers. If ``True``, the stress is automatically normalized.
+        """
+        stress_edgewise = None
+        if isinstance(stress, int) and stress >= 0:
+            stresses = self.stresses()
+            if stress >= len(stresses):
+                raise IndexError(
+                    "The value of `stress` exceeds "
+                    + "the dimension of the space "
+                    + "of equilibrium stresses."
+                )
+            stress_edgewise = self._transform_stress_to_edgewise(stresses[stress])
+        elif isinstance(stress, Matrix):
+            stress_edgewise = self._transform_stress_to_edgewise(stress)
+        elif isinstance(stress, dict) and all(
+            isinstance(stress[key], int | float | str) for key in stress.keys()
+        ):
+            stress_edgewise = stress
+        else:
+            raise TypeError("`stress` does not have the correct Type.")
 
-    @doc_category("Other")
+        if not self.is_dict_stress(stress_edgewise):
+            raise ValueError("The provided `stress` is not an equilibrium stress.")
+
+        if points is None:
+            points = self.realization(as_points=True, numerical=True)
+
+        if stress_normalization:
+            numerical_stress = {
+                edge: float(sympify(w).evalf(10)) for edge, w in stress_edgewise.items()
+            }
+            _stress = {
+                edge: round(w / np.linalg.norm(list(numerical_stress.values())), 2)
+                for edge, w in numerical_stress.items()
+            }
+        else:
+            _stress = stress_edgewise
+        if isinstance(stress_label_pos, dict):
+            if not all([self._graph.has_edge(*e) for e in stress_label_pos.keys()]):
+                raise ValueError(
+                    "The `stress_label_pos` dictionary must contain the same "
+                    + "edges as the stress dictionary."
+                )
+            for edge in self._graph.edge_list(as_tuples=True):
+                stress_keys = [set(e) for e in stress_label_pos.keys()]
+                if set(edge) not in stress_keys:
+                    stress_label_pos[edge] = 0.5
+        elif isinstance(stress_label_pos, float):
+            label_float = stress_label_pos
+            stress_label_pos = {}
+            for edge in self._graph.edge_list(as_tuples=True):
+                stress_label_pos[edge] = label_float
+        else:
+            raise TypeError(
+                "`stress_label_pos` must be either a float or a dictionary."
+            )
+        if len(points[list(points.keys())[0]]) == 2:
+            if curved_edges:
+                newGraph = nx.MultiDiGraph()
+                connection_style = self._resolve_connection_style(connection_style)
+                for e, style in connection_style.items():
+                    newGraph.add_edge(e[0], e[1], weight=style)
+                plt.box(False)  # Manually removes the frame of the plot
+                for e in newGraph.edges(data=True):
+                    edge = tuple([e[0], e[1]])
+                    nx.draw_networkx_edge_labels(
+                        newGraph,
+                        ax=ax,
+                        pos=points,
+                        edge_labels={edge: _stress[edge]},
+                        font_color=stress_color,
+                        font_size=stress_fontsize,
+                        label_pos=stress_label_pos[edge],
+                        rotate=stress_rotate_labels,
+                        connectionstyle=f"Arc3, rad = {e[2]['weight']}",
+                        **kwargs,
+                    )
+            else:
+                for edge in self._graph.edges:
+                    nx.draw_networkx_edge_labels(
+                        self._graph,
+                        ax=ax,
+                        pos=points,
+                        edge_labels={edge: _stress[edge]},
+                        font_color=stress_color,
+                        font_size=stress_fontsize,
+                        label_pos=stress_label_pos[edge],
+                        rotate=stress_rotate_labels,
+                        **kwargs,
+                    )
+        elif len(points[list(points.keys())[0]]) == 3:
+            for edge in stress_label_pos.keys():
+                pos = [
+                    points[edge[0]][i]
+                    + stress_label_pos[edge] * (points[edge[1]][i] - points[edge[0]][i])
+                    for i in range(3)
+                ]
+                ax.text(
+                    pos[0],
+                    pos[1],
+                    pos[2],
+                    str(_stress[edge]),
+                    color=stress_color,
+                    fontsize=stress_fontsize,
+                    ha="center",
+                    va="center",
+                    **kwargs,
+                )
+        else:
+            raise ValueError(
+                "The method `_plot_stress` is currently only implemented"
+                + " for frameworks in 1 up to 3 dimensions."
+            )
+
+    @doc_category("Plotting")
     def plot(
         self,
         **kwargs,
@@ -1142,7 +1539,6 @@ class Framework(object):
         ----
         Implement plotting in dimension 3 and
         """
-
         if self._dim == 3:
             return self.plot3D(**kwargs)
         elif self._dim > 3:
@@ -3303,6 +3699,131 @@ class Framework(object):
                 )
             return list(edge_order)
 
+    @doc_category("Plotting")
+    def _resolve_connection_style(self, connection_style: str) -> str:
+        """
+        Resolve the connection style for the visualization of the framework.
+
+        Parameters
+        ----------
+        connection_style:
+            The connection style for the visualization of the framework.
+        """
+        G = self._graph
+        if isinstance(connection_style, float):
+            connection_style = {
+                e: connection_style for e in G.edge_list(as_tuples=True)
+            }
+        elif isinstance(connection_style, list):
+            if not G.number_of_edges() == len(connection_style):
+                raise AttributeError(
+                    "The provided `connection_style` doesn't have the correct length."
+                )
+            connection_style = {
+                e: style
+                for e, style in zip(G.edge_list(as_tuples=True), connection_style)
+            }
+        elif isinstance(connection_style, dict):
+            if (
+                not all(
+                    [
+                        isinstance(e, tuple)
+                        and len(e) == 2
+                        and isinstance(v, float | int)
+                        for e, v in connection_style.items()
+                    ]
+                )
+                or not all(
+                    [
+                        set(key) in [set([e[0], e[1]]) for e in G.edge_list()]
+                        for key in connection_style.keys()
+                    ]
+                )
+                or any(
+                    [set(key) for key in connection_style.keys()].count(e) > 1
+                    for e in [set(key) for key in connection_style.keys()]
+                )
+            ):
+                raise AttributeError(
+                    "The provided `connection_style` contains different edges "
+                    + "than the underlying graph or has an incorrect format."
+                )
+            connection_style = {
+                e: 0
+                for e in G.edge_list(as_tuples=True)
+                if not (
+                    e in connection_style.keys()
+                    or tuple([e[1], e[0]]) in connection_style.keys()
+                )
+            } | {
+                (tuple(e) if e in G.edge_list() else tuple([e[1], e[0]])): style
+                for e, style in connection_style.items()
+            }
+        else:
+            raise TypeError(
+                "The provided `connection_style` does not have the appropriate type."
+            )
+        return connection_style
+
+    @doc_category("Plotting")
+    def _resolve_edge_colors(
+        self, edge_color: str | Sequence[Sequence[Edge]] | dict[str : Sequence[Edge]]
+    ) -> tuple[list, list]:
+        """
+        Return the lists of colors and edges in the format for plotting.
+        """
+        G = self._graph
+        edge_list = G.edge_list()
+        edge_list_ref = []
+        edge_color_array = []
+
+        if isinstance(edge_color, str):
+            return [edge_color for _ in edge_list], edge_list
+        if isinstance(edge_color, list):
+            edges_partition = edge_color
+            colors = distinctipy.get_colors(
+                len(edges_partition), colorblind_type="Deuteranomaly", pastel_factor=0.2
+            )
+            for i, part in enumerate(edges_partition):
+                for e in part:
+                    if not G.has_edge(e[0], e[1]):
+                        raise ValueError(
+                            "The input includes a pair that is not an edge."
+                        )
+                    edge_color_array.append(colors[i])
+                    edge_list_ref.append(tuple(e))
+        elif isinstance(edge_color, dict):
+            color_edges_dict = edge_color
+            for color, edges in color_edges_dict.items():
+                for e in edges:
+                    if not G.has_edge(e[0], e[1]):
+                        raise ValueError(
+                            "The input includes an edge that is not part of the framework"
+                        )
+                    edge_color_array.append(color)
+                    edge_list_ref.append(tuple(e))
+        else:
+            raise ValueError("The input color_edge has none of the supported formats.")
+        for e in edge_list:
+            if (e[0], e[1]) not in edge_list_ref and (e[1], e[0]) not in edge_list_ref:
+                edge_color_array.append("black")
+                edge_list_ref.append(e)
+        if len(edge_list_ref) > G.number_of_edges():
+            multiple_colored = [
+                e
+                for e in edge_list_ref
+                if (edge_list_ref.count(e) > 1 or (e[1], e[0]) in edge_list_ref)
+            ]
+            duplicates = []
+            for e in multiple_colored:
+                if not (e in duplicates or (e[1], e[0]) in duplicates):
+                    duplicates.append(e)
+            raise ValueError(
+                f"The color of the edges in the following list"
+                f"was specified multiple times: {duplicates}."
+            )
+        return edge_color_array, edge_list_ref
+
 
 Framework.__doc__ = Framework.__doc__.replace(
     "METHODS",
@@ -3315,6 +3836,7 @@ Framework.__doc__ = Framework.__doc__.replace(
             "Class methods",
             "Framework manipulation",
             "Infinitesimal rigidity",
+            "Plotting",
             "Other",
             "Waiting for implementation",
         ],
